Decided not to use tmux-continuum.
In general, I haven't actually gotten this to work. It always restores
something old and I forget to save, and usually a restart is a good
chance to clean out old sessions and allow myself to use a different
layout.  If I absolutely want to make sure I save/restore a session when
rebooting, I can do that manually since that's usually the exception.